tcommon/src/main/java/com/scd/common/utils/HttpUtil.java
package com.scd.common.utils;

import okhttp3.OkHttpClient;
import okhttp3.Request;
import okhttp3.Response;
import okhttp3.ResponseBody;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.io.IOException;

/**
 * @author chengdu
 * @date 2019/7/21.
 */
public class HttpUtil {

    private static final Logger LOGGER = LoggerFactory.getLogger(HttpUtil.class);

    public static String executeGet(String url) throws IOException {
        OkHttpClient client = new OkHttpClient();
        Request request = new Request.Builder()
                .url(url)
                .build();
        Response response = client.newCall(request).execute();
        LOGGER.info("response code {}", response.code());
        ResponseBody responseBody = response.body();
        return responseBody.string();
    }
}
